requirements.txt
carla==0.9.13
numpy==1.23.4
matplotlib==3.7.5
opencv-python==4.12.0.88
pygame==2.6.1
ultralytics==8.3.217
scipy==1.10.1
heapdict==1.0.1
numba==0.58.1
Pillow==10.4.0
pybind11==3.0.1
